requirements.txt
streamlit==1.36.0
onnxruntime==1.18.1
pillow==10.4.0
numpy==1.26.4